# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.FolhaPagamentoMensal:
    properties:
      ano:
        type: integer
      funcionarios:
        items:
          $ref: '#/definitions/model.FuncionarioFolhaPagamento'
        type: array
      mes:
        type: string
      total_bonificacoes:
        type: number
      total_folha:
        type: number
      total_funcionarios:
        type: integer
      total_salario_base:
        type: number
    type: object
...
  model.FuncionarioFolhaPagamento:
    properties:
      bonificacao:
        type: number
      cpf:
        type: string
      data_contratacao:
        type: string
      expediente:
        type: string
      id_funcionario:
        type: integer
      nome:
        type: string
      salario_base:
        type: number
      salario_total:
        type: number
      tipo:
        type: string
    type: object
...
  model.RelatorioFinanceiro:
    properties:
      granularity:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      projection:
        items:
          $ref: '#/definitions/model.SeriePonto'
        type: array
      series:
        items:
          $ref: '#/definitions/model.SeriePonto'
        type: array
      totals:
        properties:
          despesa:
            type: number
          lucro:
            type: number
          receita:
            type: number
        type: object
    type: object
  model.RelatorioFolhaPagamento:
    properties:
      folhas_por_mes:
        items:
          $ref: '#/definitions/model.FolhaPagamentoMensal'
        type: array
      period_end:
        type: string
      period_start:
        type: string
      tipo_filtro:
        type: string
      total_geral_folha:
        type: number
      total_periodos:
        type: integer
    type: object
  model.SeriePonto:
    properties:
      date:
        type: string
      despesa:
        type: number
      lucro:
        type: number
      receita:
        type: number
    type: object
...
  types.ErrorResponse:
    properties:
      detail:
        type: string
    type: object
info:
  contact: {}
  description: Aplicação de Banco de Dados para Gerenciamento de Bares
  title: EDNA Bar System
  version: "1.0"
paths:
...
  /relatorios/financeiro:
    get:
      consumes:
      - application/json
      description: Retrieve a financial report within a specified date range.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      - default: day
        description: Time granularity (day|week|month)
        in: query
        name: granularity
        type: string
      - description: Number of periods to project
        in: query
        name: projection_days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RelatorioFinanceiro'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Financial Report
      tags:
      - Relatórios
  /relatorios/folha-pagamento:
    get:
      consumes:
      - application/json
      description: Retrieve monthly payroll reports for a specified period. Generates
        individual payroll for each month within the date range.
      parameters:
      - description: Period start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: Period end date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      - description: Employee type filter (garcom|seguranca|caixa|faxineiro|balconista)
        in: query
        name: tipo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RelatorioFolhaPagamento'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Payroll Report
      tags:
      - Relatórios
...
swagger: "2.0"